services/dialogue_router/requirements.txt
anthropic
paho-mqtt
